Fix install-tools

+++ tools/install-tools.sh
@@ -8,10 +8,10 @@ if type apt-get > /dev/null 2>&1; then
   if ! type rsvg-convert > /dev/null 2>&1; then
     sudo apt-get install --assume-yes --no-install-recommends librsvg2-bin
   fi
-  if ! test -f "/usr/share/fonts/opentype/noto/NotoSansCJK-Regular.ttc"; then
+  if ! test -f '/usr/share/fonts/opentype/noto/NotoSansCJK-Regular.ttc'; then
     sudo apt-get install --assume-yes --no-install-recommends fonts-noto-cjk
   fi
-  if ! test -f "/usr/share/fonts/truetype/crosextra/Carlito-Regular.ttf"; then
+  if ! test -f '/usr/share/fonts/truetype/crosextra/Carlito-Regular.ttf'; then
     sudo apt-get install --assume-yes --no-install-recommends fonts-crosextra-carlito
   fi
   if ! type tlmgr > /dev/null 2>&1; then
@@ -47,9 +47,25 @@ if type apt-get > /dev/null 2>&1; then
   TLREQ=$(readlink -f "${scriptPath}/pip_requirements.txt")
   sudo env "PATH=${PATH}" pip3 --no-cache-dir install -r "${TLREQ}"
 elif type apk > /dev/null 2>&1; then
-  apk add --no-cache git curl jq librsvg font-noto-cjk
-  apk add --no-cache --repository=https://dl-cdn.alpinelinux.org/alpine/edge/community font-carlito
-  if type tlmgr > /dev/null 2>&1; then
+  if ! type git > /dev/null 2>&1; then
+    apk add --no-cache git
+  fi
+  if ! type curl > /dev/null 2>&1; then
+    apk add --no-cache curl
+  fi
+  if ! type jq > /dev/null 2>&1; then
+    apk add --no-cache jq
+  fi
+  if ! type rsvg-convert > /dev/null 2>&1; then
+    apk add --no-cache librsvg
+  fi
+  if ! test -f '/usr/share/fonts/noto/NotoSansCJK-Regular.ttc'; then
+    apk add --no-cache font-noto-cjk
+  fi
+  if ! test -f '/usr/share/fonts/carlito/Carlito-Regular.ttf'; then
+    apk add --no-cache --repository=https://dl-cdn.alpinelinux.org/alpine/edge/community font-carlito
+  fi
+  if type tlmgr > /dev/null 2>&1 && ! test -f '/opt/texlive/texdir/texmf-dist/tex/latex/lastpage/lastpage.sty'; then
     tlmgr update --self
     tlmgr install lastpage
   else
